Push pgVER behind module
